feat: positive or negative
first semester: I made this code to identify if a number is positive, negative or zero. There's a while loop running until the user types an invalid value (something other than a number), then the terminal asks if the user wants to continue and they need to choose between "sim" (yes) and "nao" (no).