feat: introduction to arrays I
Here we have an introduction to arrays. There are very basic exercises to understand how an array works in C#. There are 4 exercises inside a switch case, the first one asks the user 5 numbers and then print all of them. The second one asks the user 10 numbers and then present the sum of all of them. The third one asks the user 5 numbers and shows the largest and smallest number. And the last one asks 8 numbers and shows the average of the sequence.